backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.2.0
pypdf2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
ollama==0.1.6
requests==2.31.0
numpy==1.26.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
